alex changes again, minus config.go and some go readme changes
