+++ apps/freezoe.org/.env.production
@@ -4,4 +4,4 @@ NEXT_PUBLIC_CAMPAIGN_MAILER_API_ROOT=https://helptheducks.dxe.io
 NEXT_PUBLIC_CAMPAIGN_NAME=freezoe
 NEXT_PUBLIC_SITE_PASSWORD=
 NEXT_PUBLIC_THERMOMETER_GOAL=50000
-NEXT_PUBLIC_GOOGLE_MAPS_PLACES_NEW_API_KEY=
+NEXT_PUBLIC_GOOGLE_MAPS_PLACES_NEW_API_KEY=AIzaSyDMVfqzbB-tm8_BWleKkBJFwJaY4oLPKfw
