+++ apps/service/geocoding/google-geocoding-api.md
@@ -88,6 +88,7 @@ https://maps.googleapis.com/maps/api/geocode/json?address=95401&components=count
   "results": [],
   "status": "ZERO_RESULTS"
 }
+```
 
 ## Further Reading
